requirements.txt
opencv-python-headless>=4.8.1.78
numpy
ultralytics
supervision>=0.16
onnx
onnxruntime-gpu
numpy
tqdm
deep-sort-realtime
gradio==3.50
huggingface_hub
autodistill>=0.1.15
autodistill_grounded_sam>=0.1.1
python-dotenv
pytube